requirements.txt
dill~=0.3.5.1
FastWARC~=0.12.2
requests~=2.28.1
ftfy~=6.1.1
python-dateutil~=2.8.2
lxml~=4.9.1
dotmap~=1.3.30
more-itertools~=8.14.0
cssselect~=1.1.0